Fix listRecipesByCriteria preparationtime logic
Review: @szabo-peter

+++ backend/src/main/java/hu/flowacademy/konyhatunder/dto/SearchByCriteriaDTO.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Data
 public class SearchByCriteriaDTO {
     private String name;
-    private List<Integer> preparationTimeInterval;
+    private List<Double> preparationTimeInterval;
     private String difficulty;
     private List<String> categories;
     private Boolean hasPicture;
